requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
anthropic==0.40.0
supabase==2.9.0
ffmpeg-python==0.2.0
flask==3.0.3
apscheduler==3.10.4
